+++ backend/.env.sample
@@ -1,0 +1,3 @@
+PORT=
+JWT_SECRET=
+HASH_SECRET=
