+++ backend/model/user.model.js
@@ -1,0 +1,12 @@
+import mongoose from "mongoose";
+
+const userSchema = new mongoose.Schema({
+    name: {type: String, required: true},
+    email: {type: String, required: true, unique: true},
+    password: {type: String, required: true},
+    phoneNumber: {type: String, required: true},
+},{
+    timestamps: true
+});
+
+export const User = mongoose.model('User', userSchema);
