+++ frontend/src/Components/Footer.jsx
@@ -1,0 +1,65 @@
+import React from "react";
+import { BsTwitter } from "react-icons/bs";
+import { SiLinkedin } from "react-icons/si";
+import { BsYoutube } from "react-icons/bs";
+import { FaFacebookF } from "react-icons/fa";
+
+const Footer = () => {
+  return (
+    <div className="footer-wrapper">
+      <div className="footer-section-one">
+        <div className="footer-logo-container">
+          <h1 className="logoo">RentWize</h1>
+        </div>
+        <div className="footer-icons">
+          <a href="https://x.com/" target="_blank" rel="noopener noreferrer">
+            <BsTwitter />
+          </a>
+          <a
+            href="https://www.linkedin.com/"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <SiLinkedin />
+          </a>
+          <a
+            href="https://www.youtube.com/"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <BsYoutube />
+          </a>
+          <a
+            href="https://www.facebook.com/"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <FaFacebookF />
+          </a>
+        </div>
+      </div>
+      <div className="footer-section-two">
+        <div className="footer-section-columns">
+          <span>Qualtiy</span>
+          <span>Help</span>
+          <span>Share</span>
+          <span>Carrers</span>
+          <span>Testimonials</span>
+          <span>Work</span>
+        </div>
+        <div className="footer-section-columns">
+          <span>244-5333-7783</span>
+          <span>hello@rentwize.com</span>
+          <span>press@rentwize.com</span>
+          <span>contact@rentwize.com</span>
+        </div>
+        <div className="footer-section-columns">
+          <span>Terms & Conditions</span>
+          <span>Privacy Policy</span>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Footer;
