+++ frontend/src/assets/about-background-image.png
