+++ frontend/src/assets/about-background.png
