+++ frontend/src/assets/choose-image.png
