+++ frontend/src/assets/home-banner-background.png
